refactor: remove useless statusMessages references
