+++ src/chats/components/conversation/MessagesList.test.tsx
@@ -43,32 +43,28 @@ const user2Be: User = createMockUser({
 	id: 'user2',
 	email: 'user2@domain.com',
 	name: 'User2',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 2"
+	lastSeen: 1234567890
 });
 
 const user3Be: User = createMockUser({
 	id: 'user3',
 	email: 'user3@domain.com',
 	name: 'User3',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 3"
+	lastSeen: 1234567890
 });
 
 const user1Be: User = createMockUser({
 	id: 'user1',
 	email: 'user1@domain.com',
 	name: 'User1',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 1"
+	lastSeen: 1234567890
 });
 
 const user4Be: User = createMockUser({
 	id: 'user4',
 	email: 'user4@domain.com',
 	name: 'User4',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 4"
+	lastSeen: 1234567890
 });
 
 const room: RoomBe = {
