+++ src/chats/components/conversation/footer/MessageReferenceDisplayed.test.tsx
@@ -49,8 +49,7 @@ const user1: User = createMockUser({
 	id: 'user1',
 	email: 'user1@domain.com',
 	name: 'User 1', // on DS them will be color #FFA726
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 1"
+	lastSeen: 1234567890
 });
 
 const forwardedUser: UserBe = createMockUser({ id: 'forwardedUserId' });
