+++ src/chats/components/infoPanel/conversationActionsAccordion/ActionsAccordion.test.tsx
@@ -28,24 +28,21 @@ const user1Be: UserBe = createMockUser({
 	id: 'user1',
 	email: 'user1@domain.com',
 	name: 'User 1',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 1"
+	lastSeen: 1234567890
 });
 
 const user2Be: UserBe = createMockUser({
 	id: 'user2',
 	email: 'user2@domain.com',
 	name: 'User 2',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 2"
+	lastSeen: 1234567890
 });
 
 const user3Be: UserBe = createMockUser({
 	id: 'user3',
 	email: 'user3@domain.com',
 	name: 'User 3',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 3"
+	lastSeen: 1234567890
 });
 
 beforeEach(() => {
