+++ src/chats/components/infoPanel/conversationParticipantsAccordion/MemberList.test.tsx
@@ -10,7 +10,7 @@ import { screen, act } from '@testing-library/react';
 
 import MemberList from './MemberList';
 import useStore from '../../../../store/Store';
-import { createMockMember, createMockUser } from '../../../../tests/createMock';
+import { createMockMember, createMockRoom, createMockUser } from '../../../../tests/createMock';
 import { setup } from '../../../../tests/test-utils';
 import { RoomBe } from '../../../../types/network/models/roomBeTypes';
 import { Member, RoomType } from '../../../../types/store/RoomTypes';
@@ -20,32 +20,28 @@ const user1Be: User = createMockUser({
 	id: 'user1',
 	email: 'user1@domain.com',
 	name: 'User 1',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 1"
+	lastSeen: 1234567890
 });
 
 const user2Be: User = createMockUser({
 	id: 'user2',
 	email: 'user2@domain.com',
 	name: 'User 2',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 2"
+	lastSeen: 1234567890
 });
 
 const user3Be: User = createMockUser({
 	id: 'user3',
 	email: 'user3@domain.com',
 	name: 'User 3',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 3"
+	lastSeen: 1234567890
 });
 
 const user4Be: User = createMockUser({
 	id: 'user4',
 	email: 'user4@domain.com',
 	name: 'User 4',
-	lastSeen: 1642818617000,
-	statusMessage: "Hey there! I'm User 4"
+	lastSeen: 1642818617000
 });
 
 const user4MemberBe: Member = createMockMember({
@@ -54,81 +50,53 @@ const user4MemberBe: Member = createMockMember({
 	external: false
 });
 
-const room: RoomBe = {
+const room: RoomBe = createMockRoom({
 	id: 'Room-Id',
-	name: 'Room Name',
-	description: 'This is the description of the group',
 	type: RoomType.GROUP,
-	createdAt: '1234567',
-	updatedAt: '12345678',
-	pictureUpdatedAt: '123456789',
 	members: [
-		{
-			userId: 'user1',
-			owner: true,
-			temporary: false,
-			external: false
-		},
-		{
-			userId: 'user2',
-			owner: false,
-			temporary: false,
-			external: false
-		},
-		{
-			userId: 'user3',
-			owner: false,
-			temporary: false,
-			external: false
-		}
+		createMockMember({ userId: user1Be.id, owner: true }),
+		createMockMember({ userId: user2Be.id }),
+		createMockMember({ userId: user3Be.id })
 	]
-};
+});
+
+beforeEach(() => {
+	const store = useStore.getState();
+	store.addRoom(room);
+	store.setUserInfo(user1Be);
+	store.setUserInfo(user2Be);
+	store.setUserInfo(user3Be);
+	store.setUserInfo(user4Be);
+});
 
 describe('Participants list', () => {
 	test('The participants list should be rendered as expected', async () => {
-		const store = useStore.getState();
-		store.addRoom(room);
-		store.setUserInfo(user1Be);
-		store.setUserInfo(user2Be);
-		store.setUserInfo(user3Be);
 		setup(<MemberList roomId={room.id} />);
 		const list = await screen.findByTestId('members_list');
 		expect(list).toBeInTheDocument();
 		expect(list.children).toHaveLength(3);
-		expect(screen.getByText(/User 1/i)).toBeInTheDocument();
-		expect(screen.getByText(/User 2/i)).toBeInTheDocument();
-		expect(screen.getByText(/User 3/i)).toBeInTheDocument();
+		expect(screen.getByText(user1Be.name)).toBeInTheDocument();
+		expect(screen.getByText(user2Be.name)).toBeInTheDocument();
+		expect(screen.getByText(user3Be.name)).toBeInTheDocument();
 	});
 
 	test('Adding new user inside list should not break the list', async () => {
-		const store = useStore.getState();
-		store.addRoom(room);
-		store.setUserInfo(user1Be);
-		store.setUserInfo(user2Be);
-		store.setUserInfo(user3Be);
-		store.setUserInfo(user4Be);
 		setup(<MemberList roomId={room.id} />);
 		const list = await screen.findByTestId('members_list');
 		expect(list).toBeInTheDocument();
 		expect(list.children).toHaveLength(3);
 
-		act(() => store.addRoomMember(room.id, user4MemberBe));
+		act(() => useStore.getState().addRoomMember(room.id, user4MemberBe));
 		expect(list.children).toHaveLength(4);
 	});
 
 	test('remove user inside list should not break the list', async () => {
-		const store = useStore.getState();
-		store.addRoom(room);
-		store.setUserInfo(user1Be);
-		store.setUserInfo(user2Be);
-		store.setUserInfo(user3Be);
-		store.setUserInfo(user4Be);
 		setup(<MemberList roomId={room.id} />);
 		const list = await screen.findByTestId('members_list');
 		expect(list).toBeInTheDocument();
 		expect(list.children).toHaveLength(3);
 
-		act(() => store.removeRoomMember(room.id, 'user3'));
+		act(() => useStore.getState().removeRoomMember(room.id, user1Be.id));
 		expect(list.children).toHaveLength(2);
 	});
 
@@ -146,11 +114,6 @@ describe('Participants list', () => {
 	});
 
 	test('Search more members inside list', async () => {
-		const store = useStore.getState();
-		store.addRoom(room);
-		store.setUserInfo(user1Be);
-		store.setUserInfo(user2Be);
-		store.setUserInfo(user3Be);
 		const { user } = setup(<MemberList roomId={room.id} />);
 		const searchInput = screen.getByRole('textbox', { name: /Search members/i });
 		const list = await screen.findByTestId('members_list');
@@ -159,17 +122,12 @@ describe('Participants list', () => {
 	});
 
 	test('Search a user that is not in the list', async () => {
-		const store = useStore.getState();
-		store.addRoom(room);
-		store.setUserInfo(user1Be);
-		store.setUserInfo(user2Be);
-		store.setUserInfo(user3Be);
 		const { user } = setup(<MemberList roomId={room.id} />);
 		const searchInput = screen.getByRole('textbox', { name: /Search members/i });
 		const searchIcon = screen.getByTestId('icon: Search');
 		expect(searchIcon).toBeInTheDocument();
 		const list = await screen.findByTestId('members_list');
-		await user.type(searchInput, 'user 4');
+		await user.type(searchInput, user4Be.id);
 
 		const placeholderText = await screen.findByText(
 			/Your search returned no results, try another keyword./i
