+++ src/chats/components/secondaryBar/SecondaryBarView.test.tsx
@@ -28,24 +28,21 @@ const user1Be = createMockUser({
 	id: 'user1Id',
 	email: 'user1@domain.com',
 	name: 'User1',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 1"
+	lastSeen: 1234567890
 });
 
 const user2Be = createMockUser({
 	id: 'user2Id',
 	email: 'user2@domain.com',
 	name: 'User2',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 2"
+	lastSeen: 1234567890
 });
 
 const user3Be = createMockUser({
 	id: 'user3Id',
 	email: 'user3@domain.com',
 	name: 'User3',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 3"
+	lastSeen: 1234567890
 });
 
 const mockedGroup1: RoomBe = createMockRoom({
