+++ src/chats/components/secondaryBar/conversationList/ExpandedSidebarListItem.test.tsx
@@ -37,24 +37,21 @@ const user2Be: User = createMockUser({
 	id: 'user2Id',
 	email: 'user2@domain.com',
 	name: 'User2',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 2"
+	lastSeen: 1234567890
 });
 
 const user1Be: User = createMockUser({
 	id: 'user1Id',
 	email: 'user1@domain.com',
 	name: 'User1',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 1"
+	lastSeen: 1234567890
 });
 
 const user4Be: User = createMockUser({
 	id: 'user4Id',
 	email: 'user4@domain.com',
 	name: 'User4',
-	lastSeen: 1234567890,
-	statusMessage: "Hey there! I'm User 4"
+	lastSeen: 1234567890
 });
 
 const mockedGroup: RoomBe = createMockRoom({
