+++ src/store/slices/UsersStoreSlice.ts
@@ -9,8 +9,8 @@ import { produce } from 'immer';
 import { StateCreator } from 'zustand';
 
 import { UserBe } from '../../types/network/models/userBeTypes';
-import { RootStore, UsersStoreSlice } from '../../types/store/StoreTypes';
-import { UserType } from '../../types/store/UserTypes';
+import { RootStore } from '../../types/store/StoreTypes';
+import { UsersStoreSlice, UserType } from '../../types/store/UserTypes';
 
 export const useUsersStoreSlice: StateCreator<
 	RootStore,
@@ -28,8 +28,7 @@ export const useUsersStoreSlice: StateCreator<
 					email: user.email,
 					name: user.name,
 					type: user.type ?? UserType.INTERNAL,
-					lastSeen: user.lastSeen,
-					statusMessage: user.statusMessage
+					lastSeen: user.lastSeen
 				};
 			}),
 			false,
@@ -58,18 +57,6 @@ export const useUsersStoreSlice: StateCreator<
 			}),
 			false,
 			'USERS/SET_LAST_ACTIVITY'
-		);
-	},
-	setUserStatusMessage: (id: string, statusMsg: string): void => {
-		set(
-			produce((draft: RootStore) => {
-				draft.users[id] = {
-					...draft.users[id],
-					statusMessage: statusMsg
-				};
-			}),
-			false,
-			'USERS/SET_STATUS_MESSAGE'
 		);
 	},
 	setAnonymousUser: (id: string): void => {
