+++ src/types/network/models/userBeTypes.ts
@@ -12,5 +12,4 @@ export type UserBe = {
 	name: string;
 	type: UserType;
 	lastSeen?: number;
-	statusMessage?: string;
 };
